# swagger.json
{
  "swagger": "2.0",
  "paths": {
    "/api/users": {
      "get": {
        "description": "Fetch the registered users",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "operationId": "ListUsers",
        "responses": {
          "200": {
            "$ref": "#/responses/ListUserResponse"
          },
          "500": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "Creates a user from the request parameters",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "operationId": "CreateUserRequest",
        "parameters": [
          {
            "description": "Request to create a new User",
            "name": "CreateUserRequest",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/CreateUserResponse"
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/api/users/{userId}": {
      "get": {
        "description": "Fetches a user from the path parameter",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "operationId": "ReadUsers",
        "parameters": [
          {
            "type": "string",
            "description": "the user ID",
            "name": "userId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ReadUserResponse"
          },
          "400": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      },
      "put": {
        "description": "Updates a user from the request parameters",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "operationId": "UpdateUserRequest",
        "parameters": [
          {
            "description": "Request to update a new User",
            "name": "UpdateUserRequest",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserRequest"
            }
          },
          {
            "type": "string",
            "description": "the user ID",
            "name": "userId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/UpdateUserResponse"
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      },
      "delete": {
        "description": "Deletes a user from the path parameters",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "operationId": "ReadUsers",
        "parameters": [
          {
            "type": "string",
            "description": "the user ID",
            "name": "userId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string",
          "x-go-name": "Address"
        },
        "birthday": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "Birthday"
        },
        "created_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedAt"
        },
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "password": {
          "type": "string",
          "x-go-name": "Password"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "UpdatedAt"
        }
      },
      "x-go-package": "github.com/eijiok/user-api/model"
    },
    "UserRequest": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string",
          "x-go-name": "Address"
        },
        "birthday": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "Birthday"
        },
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "password": {
          "type": "string",
          "x-go-name": "Password"
        }
      },
      "x-go-package": "github.com/eijiok/user-api/dto"
    },
    "UserResponse": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string",
          "x-go-name": "Address"
        },
        "birthday": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "Birthday"
        },
        "created_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedAt"
        },
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "UpdatedAt"
        }
      },
      "x-go-package": "github.com/eijiok/user-api/dto"
    }
  },
  "responses": {
    "CreateUserResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/UserResponse"
      }
    },
    "ListUserResponse": {
      "description": "",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/UserResponse"
        }
      }
    },
    "ReadUserResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/UserResponse"
      }
    },
    "UpdateUserResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/UserResponse"
      }
    }
  }
}